requirements.txt
# Core data and analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
scikit-learn>=1.3.0
scipy>=1.11.0

# News and sentiment
newsapi-python>=0.2.7
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0

# Technical indicators
ta-lib>=0.4.0; platform_system != "Windows"
pandas-ta>=0.3.14b0

# Dashboard and visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Optional (for better performance)
openpyxl>=3.1.0
